Moved container_named to test_ssh_access.py
to avoid need to import conftest.py
